refactor: organization of recover password related actions
